pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "exam-analyzer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pymupdf>=1.24",
    "Pillow>=10.0",
    "pytesseract>=0.3.13",
    "anthropic>=0.40",
    "uuid6>=2024.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools]
packages = []

[dependency-groups]
dev = ["ruff>=0.8", "pytest-cov>=6.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--ignore=tests/benchmark"

[tool.pytest.ini_options.markers]
benchmark = "Real-world benchmark tests using actual PDF assets. Run with: pytest tests/benchmark/ -v -s --no-header"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", ".venv/*", "alembic/*"]
